Addressed review feedback; fixed CI issues
