+++ components/signin/public/identity_manager/BUILD.gn
@@ -11,7 +11,6 @@ source_set("identity_manager") {
   sources = [
     "brave_identity_manager.cc",
     "brave_identity_manager.h",
-    "brave_identity_manager_builder.h",
   ]
 
   public_deps = [
